1)升级ASP.NET Core 3.0和EF Core 3.0 2)修复部分异常

+++ src/Rdp.Web.Framework/Models/RoleMasterModel.cs
@@ -32,7 +32,7 @@ namespace Rdp.Web.Framework.Models
             IsValidate = 0;
             StatusFlag = 0;
         }
-        public int RoleID { get; set; }
+        public short RoleID { get; set; }
         [Required(ErrorMessageResourceType = typeof(ResMessage),
            ErrorMessageResourceName = "Required")]
         public string RoleDesc { get; set; }
